smd-ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
transformers==4.35.2
torch==2.9.1
sentence-transformers==2.2.2
nltk==3.8.1
spacy==3.7.2
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.3.2
numpy==1.25.2
# Vietnamese NLP
vncorenlp==1.0.3
underthesea==6.7.0
# LangChain for orchestration
langchain==0.1.0
langchain-community==0.0.10
# KeyBERT for keyword extraction
keybert==0.8.4
# Vietnamese OCR
vietocr==0.3.12
# Vector database
psycopg2-binary==2.9.9
pgvector==0.2.4
# Search engine
elasticsearch==8.11.0
# Additional ML models
datasets==2.14.7